feat: lazy exact warm-start weights, CMake package, Windows build through one shim
klein2 warm from its own infeasible basis tripped the iteration guard at
106201 iterations and restarted cold. A warm start guesses every
steepest-edge weight at 1.0 and each drift restarts the whole set at 1.0
again, so the walk never learned its rows. After DSE_GUESS_RESTARTS drifts
the weights are now computed exactly, one BTRAN per row; cold starts and
the tree's node solves (cfg.node_solve) are untouched, so every gate is
byte-identical. Swept 1..128 on the warm bench: 64 moves only 25fv47, at
0.25x. Exact weights at every warm start are refused in bench/refusals.txt.

CMakeLists.txt builds the archive, the shared library, the tool and the
unit suite, and installs jaosConfig.cmake beside jaos.pc; tests/cmake.sh
checks a find_package consumer.

The POSIX calls (getline, fmemopen, open_memstream, newlocale, strcasecmp,
clock_gettime) sit behind src/jaos_sys.h with a Windows half; the mingw-w64
cross build in tests/windows.sh runs jaos.exe under wine and requires the
Linux answers byte for byte.

Also: a unit test for an infeasible warm re-solve, and TODO notes the
warm bench's two pre-existing failures.

+++ src/model.c
@@ -149,8 +149,8 @@ static jaos_status name_out(jaos_model *m, const char *name, char *buf,
 {
     const size_t len = strlen(name);
     if (buf == nullptr || cap <= 0 || (size_t)cap <= len) {
-        jm_set_err(m, "the name '%s' needs %zu bytes and the buffer holds "
-                   "%lld", name, len + 1, (long long)cap);
+        jm_set_err(m, "the name '%s' needs %lld bytes and the buffer holds "
+                   "%lld", name, (long long)(len + 1), (long long)cap);
         return JAOS_ERR_INVALID_INPUT;
     }
     memcpy(buf, name, len + 1);
